requirements.txt
adsputils==v1.5.2
alembic==1.4.3
langdetect==1.0.9
nltk==3.6.7
pandas==1.3.5
psycopg2==2.8.6
regex==2022.4.24
requests==2.27.1
spacy==3.6.0
scikit_learn==1.3.0
tensorflow-cpu==2.9.1
torch==2.0.1
transformers==4.32.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.6.0/en_core_web_lg-3.6.0.tar.gz#egg=en_core_web_lg
# unidecode==1.3.7 there is a conflict with adsputils, but keeping it to make sure it is not causing error
yake @ git+https://github.com/LIAAD/yake@d2fc406c52d08843c2cb511d31d859980a7291e4
